+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Add
+- Collection.every
+- Collection.filter
 
 ## [0.0.1] - 2016-04-24
 ### Add